+++ src/codegen/expression.rs
@@ -18,6 +18,9 @@ impl Transpilable for Expression {
             Expression::Boolean(value) => String::from(if *value { "true" } else { "false" }),
             Expression::Identifier(value) => format!("{}", *value),
             Expression::FunctionCall(_) => todo!("FunctionCall codegen is not implemented yet"),
+            Expression::BinaryOperator(_, _, _) => {
+                todo!("BinaryOperator codegen is not implemented yet")
+            }
         }
     }
 }