+++ src/syntax/expression/equality.rs
@@ -1,0 +1,8 @@
+use crate::{syntax::{ParseResult, ast::Expression}, lexic::token::Token};
+
+
+pub fn try_parse(tokens: &Vec<Token>, pos: usize) -> ParseResult<Expression, ()> {
+    //
+
+    todo!()
+}